+++ stepmania/src/archutils/Unix/LinuxThreadHelpers.h
@@ -9,11 +9,11 @@ int GetCurrentThreadId();
 /* Return true if NPTL libraries are in use, false if linuxthreads. */
 bool UsingNPTL();
 
-int SuspendThread( int ThreadID );
-int ResumeThread( int ThreadID );
+int SuspendThread( uint64_t ThreadID );
+int ResumeThread( uint64_t ThreadID );
 
 struct BacktraceContext;
-int GetThreadContext( int ThreadID, BacktraceContext *ctx );
+int GetThreadContext( uint64_t ThreadID, BacktraceContext *ctx );
 	
 #endif
